♻️ refactor: 完成项目代码大规模重构与优化
✨ 主要变更：
1.  统一驱动配置类型为DriverConnectionConfig
2.  新增多个结构体的Default实现
3.  替换老旧API为现代Rust语法糖
4.  整理架构文档与代码边界
5.  优化日志与错误处理格式
6.  重构DuckDB加速查询逻辑
7.  修复多处clippy lint警告

+++ src-tauri/src/core/driver/native/mysql_pool.rs
@@ -68,7 +68,7 @@ impl DbPool for MySqlPoolWrapper {
 
     fn status(&self) -> PoolStatus {
         let size = self.pool.size() as usize;
-        let idle = self.pool.num_idle() as usize;
+        let idle = self.pool.num_idle();
         let active = size - idle;
 
         PoolStatus {
