+++ src-tauri/src/core/driver/native/postgres_pool.rs
@@ -70,7 +70,7 @@ impl DbPool for PostgresPoolWrapper {
 
     fn status(&self) -> PoolStatus {
         let size = self.pool.size() as usize;
-        let idle = self.pool.num_idle() as usize;
+        let idle = self.pool.num_idle();
         let active = size - idle;
 
         PoolStatus {
